Remove .gitignore, sample.sh, build.sh files and update README
This commit deletes the .gitignore, sample.sh, and build.sh files from the repository as they are no longer needed. The README.md file has been significantly updated with new information regarding the introduction of the MQTT functionality, integration with Home Assistant, configuration and dependencies for MQTT, and revised logic for command execution.

Signed-off-by: hldh214 <hldh214@gmail.com>